requirements.txt
sentence-transformers==2.2.2
flask==2.3.3
pypdf==3.17.1
python-dotenv==1.0.0
pinecone==6.0.2
google-generativeai==0.8.0
gunicorn==21.2.0
-e . # local package 